# docs/swagger.yaml
definitions:
...
  request.ColorRequest:
    properties:
      color_name:
        type: string
    type: object
...
  request.ProductCategoryRequest:
    properties:
      name:
        type: string
      parent_id:
        type: integer
    type: object
  request.ProductRequest:
    properties:
      brand:
        type: string
      category_id:
        type: integer
      description:
        type: string
      product_name:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      message: {}
      statusCode:
        type: integer
    type: object
  utils.UpdateCategory:
    properties:
      name:
        type: string
      parent_id:
        type: integer
    type: object
  utils.UpdateColor:
    properties:
      color_name:
        type: string
    type: object
...
host: localhost:9000
info:
  contact:
    email: engdaworkyismaw9@gmail.com
    name: Engdawork yismaw
  description: This is a sample server for an e-commerce platform.
  title: E-commerce API
  version: "1.0"
paths:
...
  /category/{id}:
    get:
      description: Get a single product category by id
      operationId: get-product-category-by-id
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get category
      tags:
      - product category
  /category/create:
    post:
      consumes:
      - application/json
      description: Insert a new product category
      operationId: create-product-category
      parameters:
      - description: Product category data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/request.ProductCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: Create category
      tags:
      - product category
  /category/delete/{id}:
    delete:
      description: Delete product category by id
      operationId: delete-product-category-by-id
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: Delete category
      tags:
      - product category
...
  /category/update/{id}:
    put:
      consumes:
      - application/json
      description: Update product category by id
      operationId: update-product-category-by-id
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update product category data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/utils.UpdateCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: Update category
      tags:
      - product category
  /color/{id}:
    get:
      description: Get a single color by id
      operationId: get-color-by-id
      parameters:
      - description: Color ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: Get color
      tags:
      - product color
  /color/create:
    post:
      consumes:
      - application/json
      description: Insert a new color
      operationId: create-color
      parameters:
      - description: Color data
        in: body
        name: color
        required: true
        schema:
          $ref: '#/definitions/request.ColorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: Create color
      tags:
      - product color
  /color/delete/{id}:
    delete:
      description: Delete color by id
      operationId: delete-color-by-id
      parameters:
      - description: Color ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: Delete color
      tags:
      - product color
  /color/list:
    get:
      description: Retrieves a list of colors
      operationId: list-color
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: List color
      tags:
      - product color
  /color/update/{id}:
    put:
      consumes:
      - application/json
      description: Update color by id
      operationId: update-color-by-id
      parameters:
      - description: Color ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update color data
        in: body
        name: color
        required: true
        schema:
          $ref: '#/definitions/utils.UpdateColor'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: Update color
      tags:
      - product color
...
  /product/create:
    post:
      consumes:
      - application/json
      description: Insert a new product
      operationId: create-product
      parameters:
      - description: Product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/request.ProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: create product
      tags:
      - product
...
schemes:
- http
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"